+++ src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import GlobalStyle from './styles/GlobalStyle';
 import { ThemeProvider } from 'styled-components';
-import Routes from './Routes';
 import theme from './styles/theme';
+import GlobalStyle from './styles/GlobalStyle';
+import Routes from './Routes';
 
 ReactDOM.render(
   <React.StrictMode>
