+++ src/pages/Auth/Auth.js
@@ -1,17 +1,23 @@
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useContext } from 'react';
+import { UserContext } from '../../contexts/UserContext';
 import { useHistory, useLocation } from 'react-router-dom';
 
 export default function Auth({ match }) {
   const { app } = match.params;
-  function useQuery() {
-    return new URLSearchParams(useLocation().search);
-  }
-  const [code, setCode] = useState(useQuery().get('code'));
-  console.log(app, code); // Test Code for Checking Variables
+  const [code, setCode] = useState(
+    new URLSearchParams(useLocation().search).get('code')
+  );
+
+  // Test Code for Checking UserContext
+  const { user, setUser } = useContext(UserContext);
+  useEffect(() => {
+    console.log(user);
+  }, [user]);
 
   const history = useHistory();
   useEffect(() => {
     if (code.length !== 0) {
+      alert('회원가입에 성공하였습니다!');
       history.push('/feed');
     } else {
       history.push('/signup');
